Refactor jobs notifications folder structure
https://eaflood.atlassian.net/browse/WATER-4961

Some of the services/ presenters related to the 'ProcessNotificationsStatusUpdatesService' are in the wrong structure.

This change updates the location of dependant services and moves them into the 'jobs/notifications' folder.
