+++ test/presenters/jobs/notifications/notify-status.presenter.test.js
@@ -8,7 +8,7 @@ const { describe, it, beforeEach } = (exports.lab = Lab.script())
 const { expect } = Code
 
 // Thing under test
-const NotifyStatusPresenter = require('../../../../app/presenters/notifications/setup/notify-status.presenter.js')
+const NotifyStatusPresenter = require('../../../../app/presenters/jobs/notifications/notify-status.presenter.js')
 
 describe('Notifications Setup - Notify status presenter', () => {
   let notifyStatus
